requirements.txt
fastapi
uvicorn
boto3
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
mangum
email-validator
